backend/requirements.txt
flask
flask-cors
pdf2docx
Pillow
pypdf
python-docx
werkzeug
docx2pdf
